requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
requests==2.28.2
rsa==4.9.1
Telethon==1.27.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
yarl==1.20.1
